upload photo function

+++ controllers/recipe_controller.js
@@ -1,4 +1,6 @@
 const { RecipeModel } = require ('../models/recipe')
+const { UserModel } = require('../models/user')
+const cloudinary = require('../config/cloudinary-config')
 
 module.exports = {
 
@@ -24,31 +26,35 @@ module.exports = {
     show: async(req, res) => {
 
         let recipeReq = {}
-        
-        RecipeModel.findOne ({_id: req.params.id})
-            .then(item => {
-                // if item is not found, redirect to homepage
-                if (!item) {
-                    res.send('Not Found')
-                    return
-                }
+       
+        let item = await RecipeModel.findOne ({_id: req.params.id})
 
-                recipeReq = item
-            })
-            .then(recipe => {
-                res.render('show', { recipe : recipeReq/*,user: user*/} )
-            })
+        if (!item) {
+            res.send('Not Found')
+            return
+        } else {
+
+            recipeReq = item
+        }
+       
+        let userReq = await UserModel.findOne({user_id: recipeReq.user_id})
+            if(userReq) {
+                res.render('show', { recipe : recipeReq ,userReq: userReq} )
+            } else {
+                res.render('show', { recipe : recipeReq } )
+            }
+            
     },
 
-    create: (req, res) => {
+    create: async (req, res) => {
 
-        const { name, cuisine, credit, serves, difficulty, time, summary, ingredient, instructions} = req.body
+        const { name, newImage, cuisine, serves, difficulty, time, summary, ingredient, instructions} = req.body
 
         //creating an errors array to display all the errors
         let errors = []
 
         //ensuring all fields are filled up
-        if(!name||  !cuisine || !serves || !difficulty || !time || !summary) {
+        if(!name|| !cuisine || !serves || !difficulty || !time || !summary) {
             errors.push('All fields are required to be filled up.')
         }
 
@@ -59,6 +65,7 @@ module.exports = {
             res.render('newform', {
                 errors: errors,
                 name: name,
+                image: newImage,
                 cuisine: cuisine,
                 serves: serves,
                 difficulty: difficulty,
@@ -80,19 +87,21 @@ module.exports = {
         
             for(let i =0; i< req.body.instruction.length; i++) {
                 newInstructions.push({"toDo": req.body.instruction[i]})
-            }      
-        
+            }
+
+            let newUpload = await cloudinary.uploader.upload(req.file.path);
 
             RecipeModel.create({
 
-                name: req.body.name,
-                cuisine: req.body.cuisine,
+                name: name,
+                image: newUpload.secure_url,
+                cloudinary_id: newUpload.public_id,
+                cuisine: cuisine,
                 user_id: req.user.user_id,
-                credit: req.body.credit,
-                serves: req.body.serves,
-                difficulty: req.body.difficulty,
-                time: req.body.time,
-                summary: req.body.summary,
+                serves: serves,
+                difficulty: difficulty,
+                time: time,
+                summary: summary,
                 ingredient: newIngredient,            
                 instruction: newInstructions
             })
@@ -102,11 +111,16 @@ module.exports = {
         }
     },
 
-    delete: (req, res) => {
-        RecipeModel.deleteOne( {_id: req.params.id})
-            .then(deleteResp => {
-            res.redirect(`/user/${req.user.user_id}/dashboard`)
-            })
+    delete: async (req, res) => {
+        
+        let deleteRecipe = await RecipeModel.findOne({_id: req.params.id})
+        if (deleteRecipe.user_id === req.user.user_id) {
+            await RecipeModel.deleteOne( {_id: req.params.id})
+            await cloudinary.uploader.destroy(deleteRecipe.cloudinary_id)
+        }
+    
+        res.redirect(`/user/${req.user.user_id}/dashboard`)
+    
     }
 
 }
